+++ sdk/storage/azure-storage-blobs/test/ut/append_blob_client_test.cpp
@@ -279,7 +279,7 @@ namespace Azure { namespace Storage { namespace Test {
 
     Azure::Storage::Blobs::ListBlobsOptions options;
     options.Prefix = blobName;
-    for (auto pageResponse = m_blobContainerClient->ListBlobs(options); pageResponse.HasMorePages();
+    for (auto pageResponse = m_blobContainerClient->ListBlobs(options); pageResponse.HasPage();
          pageResponse.MoveToNextPage())
     {
       for (const auto& blob : pageResponse.Blobs)
@@ -310,7 +310,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_TRUE(getPropertiesResult.Value.IsSealed.HasValue());
     EXPECT_TRUE(getPropertiesResult.Value.IsSealed.Value());
 
-    for (auto pageResponse = m_blobContainerClient->ListBlobs(options); pageResponse.HasMorePages();
+    for (auto pageResponse = m_blobContainerClient->ListBlobs(options); pageResponse.HasPage();
          pageResponse.MoveToNextPage())
     {
       for (const auto& blob : pageResponse.Blobs)